+++ app/encoding.go
@@ -1,14 +1,14 @@
 package app
 
 import (
+	sdkparams "github.com/cosmos/cosmos-sdk/simapp/params"
 	"github.com/cosmos/cosmos-sdk/std"
-
 	"github.com/umee-network/umee/v4/app/params"
 )
 
 // MakeEncodingConfig returns the application's encoding configuration with all
 // types and interfaces registered.
-func MakeEncodingConfig() params.EncodingConfig {
+func MakeEncodingConfig() sdkparams.EncodingConfig {
 	encodingConfig := params.MakeEncodingConfig()
 	std.RegisterLegacyAminoCodec(encodingConfig.Amino)
 	std.RegisterInterfaces(encodingConfig.InterfaceRegistry)