+++ app/params/proto.go
@@ -1,22 +1,10 @@
 package params
 
 import (
-	"github.com/cosmos/cosmos-sdk/codec"
-	"github.com/cosmos/cosmos-sdk/codec/types"
-	"github.com/cosmos/cosmos-sdk/x/auth/tx"
+	"github.com/cosmos/cosmos-sdk/simapp/params"
 )
 
 // MakeEncodingConfig creates an EncodingConfig for Amino-based tests.
-func MakeEncodingConfig() EncodingConfig {
-	amino := codec.NewLegacyAmino()
-	interfaceRegistry := types.NewInterfaceRegistry()
-	marshaler := codec.NewProtoCodec(interfaceRegistry)
-	txCfg := tx.NewTxConfig(marshaler, tx.DefaultSignModes)
-
-	return EncodingConfig{
-		InterfaceRegistry: interfaceRegistry,
-		Codec:             marshaler,
-		TxConfig:          txCfg,
-		Amino:             amino,
-	}
+func MakeEncodingConfig() params.EncodingConfig {
+	return params.MakeTestEncodingConfig()
 }