+++ tests/e2e/chain.go
@@ -8,12 +8,12 @@ import (
 	"github.com/cosmos/cosmos-sdk/crypto/keys/ed25519"
 	"github.com/cosmos/cosmos-sdk/crypto/keys/secp256k1"
 	cryptotypes "github.com/cosmos/cosmos-sdk/crypto/types"
+	sdkparams "github.com/cosmos/cosmos-sdk/simapp/params"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 	tmrand "github.com/tendermint/tendermint/libs/rand"
 
 	umeeapp "github.com/umee-network/umee/v4/app"
-	"github.com/umee-network/umee/v4/app/params"
 )
 
 const (
@@ -22,7 +22,7 @@ const (
 )
 
 var (
-	encodingConfig params.EncodingConfig
+	encodingConfig sdkparams.EncodingConfig
 	cdc            codec.Codec
 )