+++ tests/grpc/price_listener.go
@@ -3,14 +3,15 @@ package grpc
 import (
 	"fmt"
 
-	"github.com/umee-network/umee/v4/tests/grpc/client"
+	"github.com/umee-network/umee/v4/client"
+	sdkclient "github.com/umee-network/umee/v4/sdkclient"
 	oracletypes "github.com/umee-network/umee/v4/x/oracle/types"
 )
 
 func listenForPrices(
-	umeeClient *client.UmeeClient,
+	umee client.Client,
 	params oracletypes.Params,
-	chainHeight *ChainHeight,
+	chainHeight *sdkclient.ChainHeightListener,
 ) (*PriceStore, error) {
 	priceStore := NewPriceStore()
 	// Wait until the beginning of a median period
@@ -26,7 +27,7 @@ func listenForPrices(
 	for i := 0; i < int(params.MedianStampPeriod); i++ {
 		height := <-chainHeight.HeightChanged
 		if isPeriodFirstBlock(height, params.HistoricStampPeriod) {
-			exchangeRates, err := umeeClient.QueryClient.QueryExchangeRates()
+			exchangeRates, err := umee.QueryExchangeRates()
 			fmt.Printf("block %d stamp: %+v\n", height, exchangeRates)
 			if err != nil {
 				return nil, err
@@ -37,7 +38,7 @@ func listenForPrices(
 		}
 	}
 
-	medians, err := umeeClient.QueryClient.QueryMedians()
+	medians, err := umee.QueryMedians()
 	if err != nil {
 		return nil, err
 	}
@@ -52,7 +53,7 @@ func listenForPrices(
 		priceStore.medians[median.ExchangeRateTuple.Denom] = median.ExchangeRateTuple.ExchangeRate
 	}
 
-	medianDeviations, err := umeeClient.QueryClient.QueryMedianDeviations()
+	medianDeviations, err := umee.QueryMedianDeviations()
 	if err != nil {
 		return nil, err
 	}